Update: marketing feedback
