--- src/components/Marketing/CampaignDetailContent/styles.js
@@ -1,58 +1,0 @@
-import styled from 'styled-components'
-
-export const Container = styled.div``
-
-export const SelectWrapper = styled.div`
-  margin-bottom: 25px;
-
-  label {
-    font-weight: 400;
-    font-size: 14px;
-    line-height: 24px;
-    margin-bottom: 10px;
-  }
-
-  > div {
-    width: 100%;
-    border: none;
-
-    > div:first-child {
-      background-color: ${props => props.theme.colors.secundary};
-      height: 43px;
-      border: none;
-      border-radius: 7.6px;
-      font-weight: 400;
-      font-size: 14px;
-      line-height: 24px;
-      color: ${props => props.theme.colors.secundaryContrast};
-    }
-    .list {
-      background-color: ${props => props.theme.colors.secundary};
-      border: none;
-      
-      .search-bar-container {
-        padding: 10px;
-        input {
-          background-color: transparent;
-          height: 44px;
-        }
-      }
-
-      .list-wrapper {
-        > div {
-          padding: 6px 10px;
-          &:hover {
-            background-color: #f2f5f7;
-          }
-        }
-      }
-    }
-  }
-`
-
-export const Option = styled.div`
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 24px;
-  color: ${props => props.theme.colors.secundaryContrast};
-`
